+++ src/display/ST7789-SPI-240x320-Lovyann/LGFXScreenDashboardResumeEntityCPUTemperature.cpp
@@ -5,7 +5,7 @@
 
 using namespace aportela::microcontroller::utils;
 
-LGFXScreenDashboardResumeEntityCPUTemperature::LGFXScreenDashboardResumeEntityCPUTemperature(LovyanGFX *display, SourceData *sourceData, uint16_t width, uint16_t height, uint16_t xOffset, uint16_t yOffset) : LGFXScreenDashboardResumeEntity(display, sourceData, width, height, xOffset, yOffset, DASHBOARD_ITEM_TYPE_LABEL[(uint8_t)DASHBOARD_ITEM_TYPE_CPU_TEMPERATURE])
+LGFXScreenDashboardResumeEntityCPUTemperature::LGFXScreenDashboardResumeEntityCPUTemperature(LovyanGFX *display, SourceData *sourceData, uint16_t width, uint16_t height, uint16_t xOffset, uint16_t yOffset) : LGFXScreenDashboardResumeEntity(display, sourceData, width, height, xOffset, yOffset, DASHBOARD_ITEM_TYPE_LABEL[DASHBOARD_ITEM_TYPE_CPU_TEMPERATURE])
 {
     if (this->parentDisplay != nullptr)
     {
